// src/task/enum/task-date.enum.ts
export enum TaskDateEnum {
  MONTHS_18_12 = 'months-18-12',
  MONTHS_12_10 = 'months-12-10',
  MONTHS_10_8 = 'months-10-8',
  MONTHS_8_6 = 'months-8-6',
  MONTHS_6_4 = 'months-6-4',
  MONTHS_4_2 = 'months-4-2',
  MONTHS_2 = 'months-2',
  MONTH_1 = 'month-1',
  WEEKS_2 = 'weeks-2',
  WEEK = 'week',
}